feat: remove hdr plugin, introduce environment to engine.
