+++ __mocks__/three/examples/jsm/loaders/RGBELoader.js.ts
@@ -1,12 +1,14 @@
+import { Texture } from 'three';
 import { vi } from 'vitest';
-import { Texture } from '../../../../three.ts';
 
 export const RGBELoader = vi.fn(function (this: any) {
     this.loadAsync = vi.fn(async (_url: string) => {
-        const tex = new Texture();
-        // add a dispose method as HDREnvironment calls dispose on it during cleanup
-        tex.dispose = vi.fn();
-        return tex as unknown as any;
+        const tex = {
+            dispose: vi.fn(),
+            mapping: undefined,
+            colorSpace: 'srgb',
+        };
+        return tex as unknown as Texture;
     });
     return this;
 });
