+++ src/engine/index.ts
@@ -2,6 +2,7 @@ export * from './Dive.ts';
 export * from './camera/PerspectiveCamera.ts';
 export * from './clock/Clock.ts';
 export * from './renderer/Renderer.ts';
+export * from './environment/Environment.ts';
 export * from './resize/ResizeManager.ts';
 export * from './scene/Scene.ts';
 export * from './view/View.ts';
