--- src/plugins/hdr/index.ts
@@ -1,2 +1,0 @@
-export * from './src/index.ts';
-export * from './types/index.ts';
