--- src/plugins/hdr/package.json
@@ -1,6 +1,0 @@
-{
-    "name": "hdr",
-    "description": "",
-    "type": "module",
-    "private": true
-}
