--- src/plugins/hdr/src/index.ts
@@ -1,1 +1,0 @@
-export * from './HDREnvorinment.ts';
